restructure postgres_logger
