+++ lite-rpc/src/cli.rs
@@ -1,13 +1,10 @@
 use std::env;
 
-use crate::{
-    DEFAULT_FANOUT_SIZE, DEFAULT_GRPC_ADDR, DEFAULT_RETRY_TIMEOUT, DEFAULT_RPC_ADDR,
-    DEFAULT_WS_ADDR, MAX_RETRIES,
-};
+use crate::{DEFAULT_FANOUT_SIZE, DEFAULT_GRPC_ADDR, DEFAULT_RETRY_TIMEOUT, DEFAULT_RPC_ADDR, DEFAULT_WS_ADDR, MAX_RETRIES};
 use anyhow::Context;
 use clap::Parser;
 use dotenv::dotenv;
-use crate::postgres_config::PostgresSessionConfig;
+use crate::postgres_logger;
 
 #[derive(Parser, Debug, Clone)]
 #[command(author, version, about, long_about = None)]
@@ -66,7 +63,7 @@ pub struct Config {
 
     /// postgres config
     #[serde(default)]
-    pub postgres: Option<PostgresSessionConfig>,
+    pub postgres: Option<postgres_logger::PostgresSessionConfig>,
 }
 
 impl Config {
@@ -173,7 +170,7 @@ impl Config {
             .map(Some)
             .unwrap_or(config.grpc_x_token4);
 
-        config.postgres = PostgresSessionConfig::new_from_env()?.or(config.postgres);
+        config.postgres = postgres_logger::PostgresSessionConfig::new_from_env()?.or(config.postgres);
 
         Ok(config)
     }
